refactor: split map gen coast, lakes, ridges, and generator hosts (#4654)
Slice B extracts register/score/grow, lakes fill/moats, ridge walk/top-up, and factory vs generate so coastline membership is O(1) without changing Pass-2 list order.
